working on pack off module in stock management, push notification changes, and other screen changes

+++ lib/pages/dashboard/dashboard_controller.dart
@@ -488,7 +488,7 @@ class DashboardController extends GetxController
     }
   }
 
-  bool isStoreMatch(List<ProductStockInfo>? list, int tempStoreId, int qty) {
+  bool isStoreMatch(List<ProductStockInfo>? list, int tempStoreId, double qty) {
     bool match = false;
 
     if (tempStoreId == AppStorage.storeId) {
